requirements.txt
discord.py==2.3.2              # Discord bot framework
python-dotenv==1.0.1           # Environment variable management (needed for loading .env files)
Flask>=2.0.0                   # Web framework for opt-in landing page
schedule==1.2.1                # Task scheduling for daily challenges
aiohttp==3.9.5                 # Asynchronous HTTP requests (useful for webhooks or API calls)
openai-agents                  # OpenAI Agents SDK (includes built-in tracing)
psutil==5.9.8                  # System and process utilities (for monitoring bot health)
pyyaml==6.0.1                  # YAML support for config files
openai==0.28.1                 # Openai API
beautifulsoup4>=4.11.1         # HTML parsing for meme scraping
